+++ src/front/js/component/user/publicar.js
@@ -1,0 +1,90 @@
+import React, { useEffect, useContext, useState } from "react";
+import { Context } from "../../store/appContext";
+import swal from "sweetalert";
+import { useNavigate } from "react-router-dom";
+
+export const Publicar = () => {
+  const { store, actions } = useContext(Context);
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    actions.resetStoreSelectors();
+  }, []);
+
+  const handleClick = () => {};
+
+  return (
+    <>
+      {localStorage.getItem("token") ? (
+        <div
+          className="caja_publicar d-flex justify-content-center col-6"
+          style={{ height: "90vh", width: "100vh" }}
+        >
+          <div className="container col-8 rounded-3 px-0 mt-0 pb-3">
+            <div className="container rounded-top px-3 pt-3 pb-2 text-center bg-white">
+              <h4>Tu Publicación</h4>
+            </div>
+
+            <div className="container rounded-bottom pt-0 pb-4 bg-white">
+              {/* comunidad */}
+              <div className="selector mx-3 mb-3">
+                <div className="pb-2">
+                  <span className="">Comunidad Autónoma</span>
+                </div>
+                <select
+                  onChange={(e) => {
+                    actions.updateComunidad(e);
+                    localStorage.setItem("pub_comunidad", store.comunidad);
+                  }}
+                  className="form-select"
+                  aria-label="Default select example"
+                  value={store.comunidad}
+                >
+                  <option className="">todas</option>
+                  {store.listacomunidades.map((item) => {
+                    let comunidad = Object.keys(item);
+                    return (
+                      <option key={comunidad} className="">
+                        {comunidad}
+                      </option>
+                    );
+                  })}
+                </select>
+              </div>
+              {/* provincia */}
+              <div className="selector mx-3 mb-3">
+                <div className="pb-2">
+                  <span className="">Provincia</span>
+                </div>
+                <select
+                  onChange={(e) => {
+                    actions.updateProvincia(e);
+                    localStorage.setItem("pub_provincia", store.provincia);
+                  }}
+                  className="form-select"
+                  aria-label="Default select example"
+                  value={store.provincia}
+                >
+                  <option className="">todas</option>
+                  {store.listaprovincias.map((elem) => (
+                    <option key={elem} className="">
+                      {elem}
+                    </option>
+                  ))}
+                </select>
+              </div>
+            </div>
+          </div>
+        </div>
+      ) : (
+        <div
+          className="d-flex justify-content-center"
+          style={{ height: "90vh", width: "100vh" }}
+        >
+          <h5>Unauthorized...</h5>
+        </div>
+      )}
+      ;
+    </>
+  );
+};
