+++ src/front/js/layout.js
@@ -13,7 +13,6 @@ import injectContext from "./store/appContext";
 import { Navbar } from "./component/navbar";
 import { Footer } from "./component/footer";
 import { Constr } from "./pages/constr";
-import { Publicar } from "./pages/publicar";
 
 //create your first component
 const Layout = () => {
@@ -35,7 +34,6 @@ const Layout = () => {
             <Route element={<Dashboard />} path="/dashboard" />
             <Route element={<Single />} path="/single/:theid" />
             <Route element={<Constr />} path="/constr/" />
-            <Route element={<Publicar />} path="/publicar/" />
             <Route element={<h1>Not found!</h1>} />
           </Routes>
           <Footer />
